consolidate config utils
