--- server/lib/config/removed-env.js
@@ -1,3 +1,0 @@
-// This tracks environment variables that used to be supported but no longer are
-// This is to assist erroring for old config no longer supported
-module.exports = ['SQLPAD_DEBUG', 'SQLPAD_TABLE_CHART_LINKS_REQUIRE_AUTH'];
